==> 팀 프로젝트 - 멋쟁이 승우처럼/CODE/01_First_Dacon.ipynb <==
### 2021_10_15_4

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import lightgbm as lgb
from sklearn.model_selection import cross_val_score
from sklearn.metrics import mean_squared_error

total = pd.read_csv('../CSV/한국가스공사_시간별_공급량_20181231.csv', encoding='cp949')

total.head(3)

### 데이터 전처리

total["구분"].unique()

d_map = {}
for i, d in enumerate(total["구분"].unique()) :
    d_map[d] = i
total["구분"] = total["구분"].map(d_map)

total.head(10)

total["연월일"] = pd.to_datetime(total["연월일"])

total['year'] = total['연월일'].dt.year
total['month'] = total['연월일'].dt.month
total['day'] = total['연월일'].dt.day
total['weekday'] = total['연월일'].dt.weekday

train_years = [2013, 2014, 2015, 2016, 2017]
val_years = [2018]

train = total[total['year'].isin(train_years)]
val = total[total['year'].isin(val_years)]

features = ['구분', 'month', 'day', 'weekday', '시간']
train_x = train[features]
train_y = train['공급량']

val_x = val[features]
val_y = val['공급량']

### 학습

d_train = lgb.Dataset(train_x, train_y)
d_val = lgb.Dataset(val_x, val_y)

params = {
    'objective': 'regression',
    'metric':'mae',
    'seed':42
}

model = lgb.train(params, d_train, 500, d_val, verbose_eval=20, early_stopping_rounds=10)
pred = model.predict(val_x)
train_mse = mean_squared_error(val_y, pred)
print("MSE :", train_mse)

# cross_val = cross_val_score(model, val_x, val_y, cv = 5,
#                            scoring = "neg_mean_squared_error")
# print("cross validation :", cross_val)


### 추론 및 결과 제출

test = pd.read_csv('../CSV/가스공급량 수요예측 모델개발 data/test.csv')
submission = pd.read_csv('../CSV/가스공급량 수요예측 모델개발 data/sample_submission.csv')

test.head(3)

submission.head(3)

test['일자'] = test['일자|시간|구분'].str.split(' ').str[0]
test['시간'] = test['일자|시간|구분'].str.split(' ').str[1].astype(int)
test['구분'] = test['일자|시간|구분'].str.split(' ').str[2]

test['일자'] = pd.to_datetime(test['일자'])
test['year'] = test['일자'].dt.year
test['month'] = test['일자'].dt.month
test['day'] = test['일자'].dt.day
test['weekday'] = test['일자'].dt.weekday

test['구분'] = test['구분'].map(d_map)

test_x = test[features]

test_x

preds = model.predict(test_x)
submission['공급량'] = preds
# submission.to_csv('../CSV/가스공급량 수요예측 모델개발 data/baseline.csv', index = False)

def nmae(true_df, pred_df):
    target_idx = true_df.iloc[:,0]
    pred_df = pred_df[pred_df.iloc[:,0].isin(target_idx)]
    pred_df = pred_df.sort_values(by=[pred_df.columns[0]], ascending=[True])
    true_df = true_df.sort_values(by=[true_df.columns[0]], ascending=[True])
    
    true = true_df.iloc[:,1].to_numpy()
    pred = pred_df.iloc[:,1].to_numpy()
    
    score = np.mean((np.abs(true-pred))/true)
    
    return score

a = pd.DataFrame({"y" : val_y})
b = pd.DataFrame({"pred_y" : pred})

a.reset_index(drop = True, inplace = True)
a.reset_index(inplace = True)
b.reset_index(inplace = True)

print("NMAE :", nmae(a, b))
