==> 팀 프로젝트 - 멋쟁이 승우처럼/CODE/02_Data_Preprocessing.ipynb <==
### 데이터
* data : 한국가스공사 시간별 공급량 데이터
* seoul : 서울 시간별 기상 데이터
* parse_dates = ['']

import pandas as pd
import numpy as np

### 2013 ~ 2018년의 모든 시간대 총 개수 구하기

print("2013-01-01 ~ 2018-12-31 총 길이 : {}".format(len(pd.date_range(start = "2013-01-01 00:00:00", 
                                                                    end = "2018-12-31 23:00:00", freq = "1h"))))
for i in range(2013, 2019) :
    print("{}년 길이 : {}".format(i, len(pd.date_range(start = f"{i}-01-01 00:00:00", 
                                                    end = f"{i}-12-31 23:00:00", freq = "1h"))))

### 출력 결과 모든 컬럼 확인 가능하게 옵션 변경

pd.set_option("display.max_columns", None)

## 가스공급량 데이터 전처리
* parse_dates = ["연월일"] -> 시간 데이터를 변경 후 추가하기 위해 사용하지 않음

data = pd.read_csv("../CSV/한국가스공사_시간별_공급량_20181231.csv", encoding = 'cp949')

data.info()

data.head(3)

data.groupby(["연월일", "구분"]).count().head(14)

data.groupby(["구분"]).count()

data["구분"].unique()

### 가스공급량 데이터 날짜 데이터 타입 변경

data.info()

data["일시"] = data["연월일"] + " " + (data["시간"] - 1).astype(str) + ":00:00"
data["일시"] = pd.to_datetime(data["일시"])

data.head(3)

data.info()

### 일시 컬럼을 활용해 컬럼 추가
* year, month, day 컬럼 추가

data["year"] = data["일시"].dt.year
data["month"] = data["일시"].dt.month
data["day"] = data["일시"].dt.day
data["hour"] = data["일시"].dt.hour
data["weekday"] = data["일시"].dt.weekday

data.head(3)

data = data[["연월일", "시간", "일시", "year", "month", "day", "hour", "weekday", "구분", "공급량"]]

data.head(3)

data["구분"].unique()

d_map = {}
for i, d in enumerate(data["구분"].unique()) :
    d_map[d] = i
data["구분_int"] = data["구분"].map(d_map)

data = data[["연월일", "시간", "일시", "year", "month", "day", "hour", "weekday", "구분", "구분_int", "공급량"]]

data.head(3)

data.to_csv("../CSV/가스공급량.csv", index = False)

## test.csv 파일 전처리

### test.csv 파일 불러오기

test = pd.read_csv("../CSV/가스공급량 수요예측 모델개발 data/test.csv")

test.head(3)

### 컬럼 생성

test['일자'] = test['일자|시간|구분'].str.split(' ').str[0]
test['시간'] = test['일자|시간|구분'].str.split(' ').str[1].astype(int)
test['구분'] = test['일자|시간|구분'].str.split(' ').str[2]

test["일시"] = test["일자"] + " " + (test["시간"] - 1).astype(str) + ":00:00"
test["일시"] = pd.to_datetime(test["일시"])

test['year'] = test['일시'].dt.year
test['month'] = test['일시'].dt.month
test['day'] = test['일시'].dt.day
test['hour'] = test['일시'].dt.hour
test['weekday'] = test['일시'].dt.weekday

test.tail(3)

test.info()

### 구분 -> int로 변경

test['구분_int'] = test['구분'].map(d_map)

### 컬럼명, 컬럼 순서 변경

# test.columns = ['일자|시간|구분', '일자', 'hour', '구분', 'year', 'month', 'day', 'weekday']
test = test[['일자|시간|구분', '일자', "시간", "구분", "구분_int", "일시", 'year', 'month', 'day', 'hour', 'weekday']]

test.head(3)

### CSV 파일로 저장

test.to_csv("../CSV/new_test.csv", index = False)

## 서울 기상 데이터

print("2013-01-01 ~ 2018-12-31 총 길이 : {}".format(len(pd.date_range(start = "2013-01-01 00:00:00", 
                                                                    end = "2018-12-31 23:00:00", freq = "1h"))))
for i in range(2013, 2019) :
    print("{}년 길이 : {}".format(i, len(pd.date_range(start = f"{i}-01-01 00:00:00", 
                                                    end = f"{i}-12-31 23:00:00", freq = "1h"))))

### 2013년도 서울 기상 데이터 불러오기

seoul = pd.read_csv("../CSV/서울 기상 데이터/2013_seoul_weather.csv", encoding = 'cp949')

seoul.head(3)

seoul.tail(3)

### 2013년 데이터에 2014 ~ 2018년도 서울 기상 데이터 병합

for i in range(2014, 2019) :
    path = "../CSV/서울 기상 데이터/" + str(i) + "_seoul_weather.csv"
#     print(path)
    temp = pd.read_csv(path, encoding = 'cp949')
    
    seoul = pd.concat([seoul, temp], ignore_index = True)
    
    print(f"{i} temp shape : {temp.shape}")

seoul.shape

seoul.info()

### 기온(°C) 컬럼 결측치 확인

# gn.loc[gn["기온(°C)"].isnull(), "기온(°C)"]
# gn.loc[1805 : 1809, ["일시", "기온(°C)"]]
# gn.loc[8830 : 8837, ["일시", "기온(°C)"]]
seoul.loc[seoul["기온(°C)"].isnull(), "기온(°C)"]

### 17520 결측치 제거
* 17519, 17521 평균으로 대체

seoul.loc[17518 : 17522, ["일시", "기온(°C)"]]

print("17519 : {}".format(seoul.loc[17519, "기온(°C)"]))
print("17520 : {}".format(seoul.loc[17520, "기온(°C)"]))
print("17521 : {}".format(seoul.loc[17521, "기온(°C)"]))

seoul.loc[17520, ["일시", "기온(°C)"]]

temperature = np.array([seoul.loc[17519, "기온(°C)"], seoul.loc[17521, "기온(°C)"]]).mean()
seoul.loc[17520, "기온(°C)"] = round(temperature, 1)
seoul.loc[17520, "기온(°C)"]

print("17519 : {}".format(seoul.loc[17519, "기온(°C)"]))
print("17520 : {}".format(seoul.loc[17520, "기온(°C)"]))
print("17521 : {}".format(seoul.loc[17521, "기온(°C)"]))

### 41513 결측치 제거
* 41512, 41514 평균으로 대체

seoul.loc[41511 : 41515, ["일시", "기온(°C)"]]

print("41512 : {}".format(seoul.loc[41512, "기온(°C)"]))
print("41513 : {}".format(seoul.loc[41513, "기온(°C)"]))
print("41514 : {}".format(seoul.loc[41514, "기온(°C)"]))

seoul.loc[41513, ["일시", "기온(°C)"]]

temperature = np.array([seoul.loc[41512, "기온(°C)"], seoul.loc[41514, "기온(°C)"]]).mean()
seoul.loc[41513, "기온(°C)"] = round(temperature, 1)
seoul.loc[41513, "기온(°C)"]

print("41512 : {}".format(seoul.loc[41512, "기온(°C)"]))
print("41513 : {}".format(seoul.loc[41513, "기온(°C)"]))
print("41514 : {}".format(seoul.loc[41514, "기온(°C)"]))

### 41895, 41896, 41897 결측치 확인
* 41894와 41898 사이의 일정한 비율로 커지는 3개의 값 찾아 대체
* (41898 - 41894) / 4
* 위 값을 41894에 일정하게 더한 값으로 대체

seoul.loc[41892 : 41900, ["일시", "기온(°C)"]]

round((13.1 + 13.5 + 13.7 + 12.8) / 4, 1)

round((21.3 - 21.4 + 28.9 - 28.2 + 17.4 - 17.2 + 20.6 - 20.2 + 17.1 - 16.5) / 5, 1)

seoul["일시"] = pd.to_datetime(seoul["일시"])

### 41895 결측치 대체
* 14시, 15시 온도차 평균으로 대체

seoul.loc[((seoul["일시"].dt.month == 10) & (seoul["일시"].dt.day == 12) & (seoul["일시"].dt.hour == 14)) |
          ((seoul["일시"].dt.month == 10) & (seoul["일시"].dt.day == 12) & (seoul["일시"].dt.hour == 15)), ["일시", "기온(°C)"]]

a = seoul.loc[((seoul["일시"].dt.month == 10) & (seoul["일시"].dt.day == 12) & (seoul["일시"].dt.hour == 15)), "기온(°C)"].sum()
b = seoul.loc[((seoul["일시"].dt.month == 10) & (seoul["일시"].dt.day == 12) & (seoul["일시"].dt.hour == 14)), "기온(°C)"].sum()
temp_15 = round((a - b + 13.7) / 5, 1)

seoul.loc[41895, "기온(°C)"] = seoul.loc[41894, "기온(°C)"] + temp_15
seoul.loc[41895, "기온(°C)"]

### 41896 결측치 대체
* 15시, 16시 온도차 평균으로 대체

seoul.loc[((seoul["일시"].dt.month == 10) & (seoul["일시"].dt.day == 12) & (seoul["일시"].dt.hour == 15)) |
          ((seoul["일시"].dt.month == 10) & (seoul["일시"].dt.day == 12) & (seoul["일시"].dt.hour == 16)), ["일시", "기온(°C)"]]

a = seoul.loc[((seoul["일시"].dt.month == 10) & (seoul["일시"].dt.day == 12) & (seoul["일시"].dt.hour == 16)), "기온(°C)"].sum()
b = seoul.loc[((seoul["일시"].dt.month == 10) & (seoul["일시"].dt.day == 12) & (seoul["일시"].dt.hour == 15)), "기온(°C)"].sum()
temp_16 = round((a - b + 14.1) / 5, 1)
temp_16

seoul.loc[41896, "기온(°C)"] = seoul.loc[41895, "기온(°C)"] + temp_16
seoul.loc[41896, "기온(°C)"]

### 41897 결측치 대체
* 16시, 17시 온도차 평균으로 대체

seoul.loc[((seoul["일시"].dt.month == 10) & (seoul["일시"].dt.day == 12) & (seoul["일시"].dt.hour == 16)) |
          ((seoul["일시"].dt.month == 10) & (seoul["일시"].dt.day == 12) & (seoul["일시"].dt.hour == 17)), ["일시", "기온(°C)"]]

a = seoul.loc[((seoul["일시"].dt.month == 10) & (seoul["일시"].dt.day == 12) & (seoul["일시"].dt.hour == 17)), "기온(°C)"].sum()
b = seoul.loc[((seoul["일시"].dt.month == 10) & (seoul["일시"].dt.day == 12) & (seoul["일시"].dt.hour == 16)), "기온(°C)"].sum()
temp_17 = round((a - b + 13.7) / 5, 1)
temp_17

seoul.loc[41897, "기온(°C)"] = round(seoul.loc[41896, "기온(°C)"] + temp_17, 1)
seoul.loc[41897, "기온(°C)"]

seoul.loc[33133 : 33140, ["일시", "기온(°C)"]]

seoul.loc[24349 : 24356, ["일시", "기온(°C)"]]

### 결측치 재확인

seoul.loc[41892 : 41900, ["일시", "기온(°C)"]]

temp = round((seoul.loc[41894, "기온(°C)"] - seoul.loc[41898, "기온(°C)"]) / 4, 1)
avg2 = round((seoul.loc[41894, "기온(°C)"] + seoul.loc[41898, "기온(°C)"]) / 2, 1)
avg1 = round((seoul.loc[41894, "기온(°C)"] + avg2) / 2, 1)
avg3 = round((avg2 + seoul.loc[41898, "기온(°C)"]) / 2, 1)
print("temp : {}, avg1 : {}, avg2 : {}, avg3 : {}".format(temp, avg1, avg2, avg3))

### 51810 결측치 제거
* 51809, 51811 평균으로 대체

seoul.loc[51808 : 51812, ["일시", "기온(°C)"]]

seoul.loc[51808, ["일시", "기온(°C)"]]

print("51809 : {}".format(seoul.loc[51809, "기온(°C)"]))
print("51810 : {}".format(seoul.loc[51810, "기온(°C)"]))
print("51811 : {}".format(seoul.loc[51811, "기온(°C)"]))

seoul.loc[51810, "기온(°C)"] = np.array([seoul.loc[51809, "기온(°C)"], seoul.loc[51811, "기온(°C)"]]).mean()
seoul.loc[51810, "기온(°C)"]

# (seoul.loc[51809, "기온(°C)"] + seoul.loc[51811, "기온(°C)"]) / 2
# np.mean([[seoul.loc[51809, "기온(°C)"]], [seoul.loc[51811, "기온(°C)"]]])
# arr = np.array([seoul.loc[51809, "기온(°C)"], seoul.loc[51811, "기온(°C)"]])
# arr.mean()
# np.array([seoul.loc[51809, "기온(°C)"], seoul.loc[51811, "기온(°C)"]]).mean()

print("51809 : {}".format(seoul.loc[51809, "기온(°C)"]))
print("51810 : {}".format(seoul.loc[51810, "기온(°C)"]))
print("51811 : {}".format(seoul.loc[51811, "기온(°C)"]))

### 기온(°C) 컬럼 결측치 재확인

seoul.loc[seoul["기온(°C)"].isnull(), "기온(°C)"]

seoul.info()

### 기온 결측치 제거된 파일 CSV 저장

seoul.to_csv("../CSV/서울 기상 데이터/2013_2018_seoul_weather.csv", index = False)

### 서울 기온 데이터 추출
* 서울 기상 데이터에서 일자와 기온 컬럼만 가지는 데이터 생성
* 일자 데이터를 통해 새로운 컬럼 생성
    * year
    * month
    * day
    * hour
    * weekday

seoul_temperature = seoul[["일시", "기온(°C)"]]

seoul_temperature.info()

### 새로운 컬럼 생성
* year
* month
* day
* hour
* weekday

seoul_temperature["year"] = seoul_temperature["일시"].dt.year
seoul_temperature["month"] = seoul_temperature["일시"].dt.month
seoul_temperature["day"] = seoul_temperature["일시"].dt.day
seoul_temperature["hour"] = seoul_temperature["일시"].dt.hour
seoul_temperature["weekday"] = seoul_temperature["일시"].dt.weekday

seoul_temperature.head(3)

seoul_temperature.to_csv("../CSV/서울 기상 데이터/2013_2018_seoul_temperature.csv", index = False)
