==> 팀 프로젝트 - 멋쟁이 승우처럼/CODE/03_Data_Merge.ipynb <==
import pandas as pd

print("2013-01-01 ~ 2018-12-31 총 길이 : {}".format(len(pd.date_range(start = "2013-01-01 00:00:00", 
                                                                    end = "2018-12-31 23:00:00", freq = "1h"))))
for i in range(2013, 2019) :
    print("{}년 길이 : {}".format(i, len(pd.date_range(start = f"{i}-01-01 00:00:00", 
                                                    end = f"{i}-12-31 23:00:00", freq = "1h"))))

gas_data = pd.read_csv("../CSV/가스공급량.csv", parse_dates = ["일시"])
new_test = pd.read_csv("../CSV/new_test.csv", parse_dates = ["일시"])
seoul_temperature = pd.read_csv("../CSV/서울 기상 데이터/2013_2018_seoul_temperature.csv", parse_dates = ["일시"])

gas_data.head(3)

seoul_temperature.tail(3)

new_test.head(3)

gas_data.tail(3)

temp = pd.DataFrame()

temp.shape

# for i in range(6) :
#     temp = pd.concat([temp, seoul_temperature])
#     print("{} 번째".format(i + 1), end = " - ")
#     print("temp shape : ", temp.shape)

# print("gas_data shape : ", gas_data.shape)

seoul_temperature[seoul_temperature["year"] == 2013]

for i in range(2013, 2019, 1) :
    for j in range(7) :
        temp = pd.concat([temp, seoul_temperature[seoul_temperature["year"] == i]])
        print(j + 1, end = " ")
    print("{} 년도".format(i), end = " - ")
    print("temp shape :", temp.shape)

temp.shape

gas_data.shape

temp.tail(3)

temp.reset_index(drop = True, inplace = True)

temp.tail(3)

gas_data["기온(°C)"] = temp["기온(°C)"]

gas_data.tail(10)

seoul_temperature.tail(10)

### 기온 데이터 추가한 가스공급량 데이터 저장

gas_data.to_csv("../CSV/train_data.csv", index = False)

gas_data.loc[122638:122642, :]

seoul_temperature.loc[17518 : 17522, :]

seoul_temperature.loc[8758:8762, :]

8760 * 7 + 8760 * 7

8760 * 2

gas_data.shape
