==> 팀 프로젝트 - 멋쟁이 승우처럼/CODE/08_Gas_Supply_Prediction_Model.ipynb <==
import pandas as pd
import numpy as np
import time

from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_absolute_error
from sklearn.metrics import mean_squared_error

from sklearn.linear_model import LinearRegression
from sklearn.tree import DecisionTreeRegressor 
from sklearn.ensemble import RandomForestRegressor   # 앙상블(의사결정트리 확장판)
from sklearn.ensemble import GradientBoostingRegressor

import xgboost as xgb
import lightgbm as lgb

train = pd.read_csv("../CSV/train_data.csv")
test = pd.read_csv("../CSV/pred_test.csv")

print("train shape : {}, test shape : {}".format(train.shape, test.shape))

train.head(3)

### LinearRegression

X = train[["year", "month", "day", "hour", "weekday", "기온(°C)", "구분_int"]]
y = train["공급량"]

for i in range(1, 10, 1) :
    X_train, X_test, y_train, y_test = train_test_split(X, y,
                                                       test_size = i / 10,
                                                       random_state = 77)

    lr_model = LinearRegression()
    lr_model.fit(X_train, y_train)
    pred = lr_model.predict(X_test)

    # 결정계수 확인
    print("학습용 : {}, 테스트용 : {}".format(10 - i, i))
    print("학습용 데이터 결정계수: {:.3f}".format(lr_model.score(X_train, y_train)))
    print("테스트 데이터 결정계수: {:.3f}".format(lr_model.score(X_test, y_test)))
    print("MSE :", mean_squared_error(y_test, pred))
    print()
    
# model.feature_importances_

### xgboost

X = train[["year", "month", "day", "hour", "weekday", "기온(°C)", "구분_int"]]
y = train["공급량"]

for i in range(1, 10, 1) :
    X_train, X_test, y_train, y_test = train_test_split(X, y,
                                                       test_size = i / 10,
                                                       random_state = 77)

    xgb_model = xgb.XGBRegressor(objective = "reg:linear", 
                            colsample_bytree = 0.3, 
                            learning_rate = 0.1, 
                            max_depth = 3, 
                            alpha = 0.1, 
                            n_estimators = 1000)
    xgb_model.fit(X_train, y_train)
    pred = xgb_model.predict(X_test)

    # 결정계수 확인
    print("학습용 : {}, 테스트용 : {}".format(10 - i, i))
    print("학습용 데이터 결정계수: {:.3f}".format(xgb_model.score(X_train, y_train)))
    print("테스트 데이터 결정계수: {:.3f}".format(xgb_model.score(X_test, y_test)))
    print("MSE :", mean_squared_error(y_test, pred))
    print()

### lightgbm

X = train[["year", "month", "day", "hour", "weekday", "기온(°C)", "구분_int"]]
y = train["공급량"]

for i in range(1, 10, 1) :
    X_train, X_test, y_train, y_test = train_test_split(X, y,
                                                       test_size = i / 10,
                                                       random_state = 77)
    
    hyperparameters = {'boosting_type': 'gbdt', 
                       'colsample_bytree': 0.7250136792694301, 
                       'is_unbalance': False, 
                       'learning_rate': 0.013227664889528229,
                       'min_child_samples': 20, 
                       'num_leaves': 56, 
                       'reg_alpha': 0.7543896477745794, 
                       'reg_lambda': 0.07152751159655985, 
                       'subsample_for_bin': 240000, 
                       'subsample': 0.5233384321711397, 
                       'n_estimators': 2000}

    lgb_model = lgb.LGBMRegressor(**hyperparameters)
    lgb_model.fit(X_train, y_train)
    pred = lgb_model.predict(X_test)

    true_y = pd.DataFrame({"true_y" : y_test})
    pred_y = pd.DataFrame({"pred_y" : pred})

    true_y.reset_index(drop = True, inplace = True)
    true_y.reset_index(inplace = True)
    pred_y.reset_index(inplace = True)
    
    # 결정계수 확인
    print("학습용 : {}, 테스트용 : {}".format(10 - i, i))
    print("학습용 데이터 결정계수: {:.3f}".format(lgb_model.score(X_train, y_train)))
    print("테스트 데이터 결정계수: {:.3f}".format(lgb_model.score(X_test, y_test)))
    print("MAE : {:.6f}".format(mean_absolute_error(y_test, pred)))
    print("MSE : {:.6f}".format(mean_squared_error(y_test, pred)))
    print("RMSE : {:.6f}".format(np.sqrt(mean_squared_error(y_test, pred))))
    print("NMAE : {:.6f}".format(nmae(true_y, pred_y)))
    print()

### 다양한 모델 비교
* LinearRegression
* DecisionTreeRegressor
* RandomForestRegressor
* GradientBoostingRegressor

def nmae(true_df, pred_df):
    target_idx = true_df.iloc[:,0]
    pred_df = pred_df[pred_df.iloc[:,0].isin(target_idx)]
    pred_df = pred_df.sort_values(by=[pred_df.columns[0]], ascending=[True])
    true_df = true_df.sort_values(by=[true_df.columns[0]], ascending=[True])
    
    true = true_df.iloc[:,1].to_numpy()
    pred = pred_df.iloc[:,1].to_numpy()
    
    score = np.mean((np.abs(true-pred))/true)
    
    return score

start_time = time.time()

X = train[["year", "month", "day", "hour", "weekday", "구분_int", "기온(°C)"]]
y = train["공급량"]

model_list = ["LinearRegression", 
              "DecisionTreeRegressor", 
              "RandomForestRegressor", 
              "GradientBoostingRegressor", 
              "xgboost", 
              "lightgbm"]

models = [LinearRegression(), 
         DecisionTreeRegressor(), 
         RandomForestRegressor(n_jobs = -1, random_state = 37), 
         GradientBoostingRegressor(random_state = 37), 
         xgb.XGBRegressor(), 
         lgb.LGBMRegressor()]

test_size = []
train_score = []
test_score = []
MAE = []
MSE = []
RMSE = []
NMAE = []
model_name = []
    
for idx, model in enumerate(models) :
    for i in range(1, 10, 1) :
        X_train, X_test, y_train, y_test = train_test_split(X, y,
                                                           test_size = i / 10,
                                                           random_state = 77)
        print("model :", model_list[idx])
        model.fit(X_train, y_train)
        pred = model.predict(X_test)
        
        true_y = pd.DataFrame({"true_y" : y_test})
        pred_y = pd.DataFrame({"pred_y" : pred})
        
        true_y.reset_index(drop = True, inplace = True)
        true_y.reset_index(inplace = True)
        pred_y.reset_index(inplace = True)

        # 결정계수 확인
        print("학습용 : {}, 테스트용 : {}".format(10 - i, i))
        print("학습용 데이터 결정계수: {:.3f}".format(model.score(X_train, y_train)))
        print("테스트 데이터 결정계수: {:.3f}".format(model.score(X_test, y_test)))

        # 평가 지표
        print("MAE : {:.6f}".format(mean_absolute_error(y_test, pred)))
        print("MSE : {:.6f}".format(mean_squared_error(y_test, pred)))
        print("RMSE : {:.6f}".format(np.sqrt(mean_squared_error(y_test, pred))))
        print("NMAE : {:.6f}".format(nmae(true_y, pred_y)))
        print()
        test_size.append(i)
        train_score.append(model.score(X_train, y_train))
        test_score.append(model.score(X_test, y_test))
        MAE.append(mean_absolute_error(y_test, pred))
        MSE.append(mean_squared_error(y_test, pred))
        RMSE.append(np.sqrt(mean_squared_error(y_test, pred)))
        NMAE.append(nmae(true_y, pred_y))
        model_name.append(model_list[idx])
        
print("실행 시간 : {:.3f}".format(time.time() - start_time))

print("test_size length : ", len(test_size))
print("train_score length :", len(train_score))
print("test_score length :", len(test_score))
print("MAE length :", len(MAE))
print("MSE length :", len(MSE))
print("RMSE length :", len(RMSE))
print("NMAE length :", len(NMAE))
print("model_name length :", len(model_name))

data_dict = {"model" : model_name, "test_size" : test_size, "train_score" : train_score, 
            "test_score" : test_score, "MAE" : MAE, "MSE" : MSE, "RMSE" : RMSE, "NMAE" : NMAE}
df = pd.DataFrame(data_dict)

df["train_score"] = df["train_score"] * 100
df["test_score"] = df["test_score"] * 100
df

df.describe()

### 최종 모델 선택

X = train[["year", "month", "day", "hour", "weekday", "구분_int", "기온(°C)"]]
y = train["공급량"]

i = 2

X_train, X_test, y_train, y_test = train_test_split(X, y,
                                                   test_size = i / 10,
                                                   random_state = 77)

model = RandomForestRegressor(n_jobs = -1, random_state = 37)
model.fit(X_train, y_train)
pred = model.predict(X_test)

true_y = pd.DataFrame({"true_y" : y_test})
pred_y = pd.DataFrame({"pred_y" : pred})

true_y.reset_index(drop = True, inplace = True)
true_y.reset_index(inplace = True)
pred_y.reset_index(inplace = True)
        
# 결정계수 확인
print("학습용 : {}, 테스트용 : {}".format(10 - i, i))
print("학습용 데이터 결정계수: {:.3f}".format(model.score(X_train, y_train)))
print("테스트 데이터 결정계수: {:.3f}".format(model.score(X_test, y_test)))

# 평가 지표
print("MAE : {:.6f}".format(mean_absolute_error(y_test, pred)))
print("MSE : {:.6f}".format(mean_squared_error(y_test, pred)))
print("RMSE : {:.6f}".format(np.sqrt(mean_squared_error(y_test, pred))))
print("NMAE : {:.6f}".format(nmae(true_y, pred_y)))
    
# model.feature_importances_

submission = pd.read_csv('../CSV/가스공급량 수요예측 모델개발 data/sample_submission.csv')

submission.head(3)

test_x = test[["year", "month", "day", "hour", "weekday", "구분_int", "기온(°C)"]]
pred = model.predict(test_x)
submission["공급량"] = pred

pred

submission.head(3)

submission.tail(3)

submission.to_csv("../CSV/Submission/05_7_columns_RandomForest_model.csv", index = False)

start_time = time.time()

X = train[["year", "month", "day", "hour", "weekday", "구분_int", "기온(°C)"]]
y = train["공급량"]

test_size = []
train_score = []
test_score = []
MAE = []
MSE = []
RMSE = []
NMAE = []
model_name = []

for i in range(1, 10, 1) :
    X_train, X_test, y_train, y_test = train_test_split(X, y,
                                                       test_size = i / 10,
                                                       random_state = 77)
    d_train = lgb.Dataset(X_train, y_train)
    d_val = lgb.Dataset(X_test, y_test)
    
    params = {}
    params["objective"] = "regression"
    params["boosting_type"] = "dart"
    params["metric"] = "mse"
    params["seed"] = 42
#     params["learning_rate"] = 0.003
#     params['sub_feature'] = 0.5
#     params['num_leaves'] = 10
#     params['min_data'] = 50
#     params['max_depth'] = 10

    model = lgb.train(params, d_train, 500, d_val, verbose_eval=20, early_stopping_rounds=10)
#     model.fit(X_train, y_train)
    pred = model.predict(X_test)

    true_y = pd.DataFrame({"true_y" : y_test})
    pred_y = pd.DataFrame({"pred_y" : pred})

    true_y.reset_index(drop = True, inplace = True)
    true_y.reset_index(inplace = True)
    pred_y.reset_index(inplace = True)

    # 결정계수 확인
    print("학습용 : {}, 테스트용 : {}".format(10 - i, i))
#     print("학습용 데이터 결정계수: {:.3f}".format(model.score(X_train, y_train)))
#     print("테스트 데이터 결정계수: {:.3f}".format(model.score(X_test, y_test)))

    # 평가 지표
    print("MAE : {:.6f}".format(mean_absolute_error(y_test, pred)))
    print("MSE : {:.6f}".format(mean_squared_error(y_test, pred)))
    print("RMSE : {:.6f}".format(np.sqrt(mean_squared_error(y_test, pred))))
    print("NMAE : {:.6f}".format(nmae(true_y, pred_y)))
    print()
    test_size.append(i)
#     train_score.append(model.score(X_train, y_train))
#     test_score.append(model.score(X_test, y_test))
    MAE.append(mean_absolute_error(y_test, pred))
    MSE.append(mean_squared_error(y_test, pred))
    RMSE.append(np.sqrt(mean_squared_error(y_test, pred)))
    NMAE.append(nmae(true_y, pred_y))
        
print("실행 시간 : {:.3f}".format(time.time() - start_time))

print("test_size length : ", len(test_size))
print("MAE length :", len(MAE))
print("MSE length :", len(MSE))
print("RMSE length :", len(RMSE))
print("NMAE length :", len(NMAE))

data_dict = {"test_size" : test_size, "MAE" : MAE, "MSE" : MSE, "RMSE" : RMSE, "NMAE" : NMAE}
gbdt_df = pd.DataFrame(data_dict)
gbdt_df

start_time = time.time()

X = train[["year", "month", "day", "hour", "weekday", "구분_int", "기온(°C)"]]
y = train["공급량"]

model_list = ["LinearRegression", 
              "DecisionTreeRegressor", 
              "RandomForestRegressor", 
              "GradientBoostingRegressor", 
              "xgboost", 
              "lightgbm"]

models = [LinearRegression(), 
         DecisionTreeRegressor(), 
         RandomForestRegressor(n_jobs = -1, random_state = 37), 
         GradientBoostingRegressor(random_state = 37), 
         xgb.XGBRegressor(), 
         lgb.LGBMRegressor()]

train_score = []
test_score = []
MAE = []
MSE = []
RMSE = []
NMAE = []
model_name = []
    
for idx, model in enumerate(models) :
    X_train, X_test, y_train, y_test = train_test_split(X, y, random_state = 77)
    print("model :", model_list[idx])
    model.fit(X_train, y_train)
    pred = model.predict(X_test)

    true_y = pd.DataFrame({"true_y" : y_test})
    pred_y = pd.DataFrame({"pred_y" : pred})

    true_y.reset_index(drop = True, inplace = True)
    true_y.reset_index(inplace = True)
    pred_y.reset_index(inplace = True)

    # 결정계수 확인
    print("학습용 데이터 결정계수: {:.3f}".format(model.score(X_train, y_train)))
    print("테스트 데이터 결정계수: {:.3f}".format(model.score(X_test, y_test)))

    # 평가 지표
    print("MAE : {:.6f}".format(mean_absolute_error(y_test, pred)))
    print("MSE : {:.6f}".format(mean_squared_error(y_test, pred)))
    print("RMSE : {:.6f}".format(np.sqrt(mean_squared_error(y_test, pred))))
    print("NMAE : {:.6f}".format(nmae(true_y, pred_y)))
    print()
    train_score.append(np.round(model.score(X_train, y_train), 6))
    test_score.append(np.round(model.score(X_test, y_test), 6))
    MAE.append(np.round(mean_absolute_error(y_test, pred), 6))
    MSE.append(np.round(mean_squared_error(y_test, pred), 6))
    RMSE.append(np.round(np.sqrt(mean_squared_error(y_test, pred)), 6))
    NMAE.append(np.round(nmae(true_y, pred_y), 6))
    model_name.append(model_list[idx])
        
print("실행 시간 : {:.3f}".format(time.time() - start_time))

print("train_score length :", len(train_score))
print("test_score length :", len(test_score))
print("MAE length :", len(MAE))
print("MSE length :", len(MSE))
print("RMSE length :", len(RMSE))
print("NMAE length :", len(NMAE))
print("model_name length :", len(model_name))

data_dict = {"model" : model_name, "train_score" : train_score, 
            "test_score" : test_score, "MAE" : MAE, "MSE" : MSE, "RMSE" : RMSE, "NMAE" : NMAE}
df = pd.DataFrame(data_dict)

df["train_score"] = df["train_score"] * 100
df["test_score"] = df["test_score"] * 100
df

df.to_csv("../CSV/model_compare.csv", index = False)
