==> 팀 프로젝트 - 멋쟁이 승우처럼/CODE/16_Submission_Merge.ipynb <==
import pandas as pd
import numpy as np

submission = pd.read_csv('../CSV/가스공급량 수요예측 모델개발 data/sample_submission.csv')

df_1 = pd.read_csv("../CSV/Submission/17_4_features_DeepLearning_model.csv")
df_2 = pd.read_csv("../CSV/Submission/18_4_features_DeepLearning_model.csv")
df_3 = pd.read_csv("../CSV/Submission/25_4_features_DeepLearning_model.csv")

df_1.head(3)

submission['공급량'] = (df_1["공급량"] + df_2["공급량"] + df_3["공급량"]) / 3
submission.head(3)

submission.to_csv("../CSV/Submission/31_4_features_DeepLearning_model.csv", index = False)
